Line 98: Added tagGroup.
Acts the same as the S link but for any tag. Fully serves its purpose here when used on a master tag with slaves and/or aliases. If used on a slave, will act the same as the slave link at index 1 and redirect to the corresponding master (redundancy alert).